feat: provide your own mapbox library/geocoder
In scenarios where you deploy your own web app, with sveltekit or not,
using the window object is fine, since you control everything about the page.
However, in browser extensions, we can't expect to populate the window object,
since we might collide with the underlying page or other extensions.

This PR allows you to Bring Your Own Mapbox, by providing the library directly
to the component.